Ensure Context is not dellocated while C code is still using it
As far as Go is concerned, Context.ctx and Context.cbc are just values, so it is
OK to implement
> cgo_call(c.ctx)
as
> tmp := c.ctx
> dropReferenceTo(c)
> // c's finalizer can be called at any point from now on, possibly deallocating c.ctx
> cgo_call(tmp) // possibly using free memory

To fix this, explicitly extend the lifetime of Context:
> cgo_call(c.ctx)
> runtime.KeepAlive(c)
becomes
> tmp := c.ctx
> cgo_call(tmp) // c.ctx is certainly still valid
> dropReferenceTo(c)
> // c's finalizer can be called from now on, possibly deallocating c.ctx

To be extra careful, this adds runtime.KeepAlive calls even in cases where the current
code path necessarily keeps c alive afterwards, so that the patern is clear and so
that future possible changes to the code don't invalidate this assumption.

Secondarily, in a few cases GPGMe returns internal pointers inside Context.ctx;
in that case, extend the lifetime of Context until we make a complete copy of
the data.

As a special case, EngineInfo is now not just a wrapper around a C pointer with
an unknown lifetime, but we make a full copy of the data when obtaining the pointer
so that we can safely use it after dropping the context that returned the information,
or after changing the information and possibly invalidating the pointers.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>