Update from very old GnuPG 1.4
The last release was 1.4.23 in Jun 2018; it's not very
relevant in modern environments.

More importantly, the unit tests have been consistently failing
on macOS recently, and per https://github.com/proglottis/gpgme/pull/48
it seems that the underlying GnuPG does not provide the expected output,
and updating GnuPG fixes that.

We've been pinned to 1.x for ctx.SetCallback, but that feature
now exists in 2.x as well, as long as callers opt in. So, do so.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>